+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+
+<a name="v2.0.2"></a>
+## [v2.0.2] - 2021-12-08
 ### Conf
 - **nvim-tree:** nvim tree highlight set to 3
 
@@ -250,7 +253,8 @@
 - **indentation:** moved indentation to it's own folder
 
 
-[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.0.2...HEAD
+[v2.0.2]: https://github.com/tigorlazuardi/nvim/compare/v2.0.1...v2.0.2
 [v2.0.1]: https://github.com/tigorlazuardi/nvim/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/tigorlazuardi/nvim/compare/v1.3.1...v2.0.0
 [v1.3.1]: https://github.com/tigorlazuardi/nvim/compare/v1.3.0...v1.3.1